fixed malloc incorrect sizeof type

+++ lab_03/exercise_3/src/thread_routines.c
@@ -83,7 +83,8 @@ void startThreads() {
 
   // Processed buffer stuff
   int processed_buffer_size = input_buffer_size;
-  double *processed_buffer = (int *)malloc(processed_buffer_size * sizeof(int));
+  double *processed_buffer =
+      (double *)malloc(processed_buffer_size * sizeof(double));
   thread_data.processed_buffer = processed_buffer;
   thread_data.processed_buffer_size = processed_buffer_size;
   thread_data.n_processed_values = 0;
